adding screenshots for debugging connection e2e tests
